+++ components/service/variantinfo/variantInfoSources/myVariantInfo.tsx
@@ -1,8 +1,5 @@
 import { MetaNode } from '@/spec/metanode'
-import { VariantTerm } from '@/components/core/term'
 import { z } from 'zod'
-import { resolveVariantCaID, variantIdResolveErrorMessage, alleleRegRespErrorMessage, getMyVarintInfoLink } from '../variantUtils'
-import { getAlleleRegistryVariantInfo } from './alleleRegistryVariantInfo'
 
   export const MyVariantInfoC = z.object({
     '_id':z.string()
@@ -29,34 +26,3 @@ import { getAlleleRegistryVariantInfo } from './alleleRegistryVariantInfo'
     </div>
   ))
   .build()
-
-  
-  export const VariantInfoFromVariantTermMyVarintInfo = MetaNode('VariantInfoFromVariantTermMyVarintInfo')
-  .meta({
-    label: 'Resolve Variant Info from Term (MyVarintInfo)',
-    description: 'Resolve variant info from variant term using the MyVarintInfo API.',
-  })
-  .inputs({ variant: VariantTerm })
-  .output(MyVariantInfo)
-  .resolve(async (props) => {
-    var varCaId = await resolveVariantCaID(props.inputs.variant);
-    if(varCaId == null || varCaId == ''){
-      throw new Error(variantIdResolveErrorMessage);
-    }
-
-    const alleleRegResponse = await getAlleleRegistryVariantInfo(varCaId);
-    if(alleleRegResponse == null){
-      throw new Error(alleleRegRespErrorMessage);
-    }
-
-    let myVariantInfoURL = getMyVarintInfoLink(alleleRegResponse);
-
-    if(myVariantInfoURL != null){
-      return await getVariantInfoFromMyVariantInfo(myVariantInfoURL);
-    }else{
-      throw new Error("Unable to find requested data, missing MyVarintInfo API link in External resources!");
-    }
-  })
-  .story(props => ({}))
-  .build()
-
